+++ sources/PieceSwap/PieceSwapScript.move
@@ -53,7 +53,7 @@ module piece_swap_script {
         assert!(!coin_list::is_coin_in_list<piece_swap::LPToken<Y,X>>(admin_addr), E_LP_TOKEN_ALREADY_IN_COIN_LIST);
 
         let decimals = math::max((coin::decimals<X>() as u128), (coin::decimals<Y>() as u128));
-        let decimals = (decimals as u64);
+        let decimals = (decimals as u8);
 
         piece_swap::create_new_pool<X, Y>(
             admin,
